requirements.txt
google-generativeai
python-dotenv
pandas
tqdm
